requirements.txt
fastapi
jinja2
uvicorn
transmission-rpc
python-dotenv
paramiko